Refactor: -  fixing jnb
