+++ src/konnektor/tests/test_data.py
@@ -1,10 +1,16 @@
 
 import pytest
 from gufe import SmallMoleculeComponent
-from konnektor.data import get_benzene_compounds
+from konnektor.data import get_benzene_ligands, get_hif2a_ligands
 
 def test_get_benzenes():
-    compounds = get_benzene_compounds()
+    compounds = get_benzene_ligands()
 
     assert len(compounds) == 16
+    assert all(isinstance(c, SmallMoleculeComponent) for c in compounds)
+
+def test_get_hif2a():
+    compounds = get_hif2a_ligands()
+
+    assert len(compounds) == 37
     assert all(isinstance(c, SmallMoleculeComponent) for c in compounds)
